refactor(buildrequest): extract fallback chain resolution into fallback_resolve.go
Move Concern G in full from orchestrator.go to a new
fallback_resolve.go (~280 LOC). Verbatim move with no behaviour change;
declarations follow the original order:
ResolveFallbackChainWithReason, ResolveFallbackChainForClientWithReason
(G1) and the reason-discarding wrappers ResolveFallbackChain,
ResolveFallbackChainForClient (G2). Imports: bamlutils,
buildrequest/roundrobin (for InspectStrategyOverride and
IsRoundRobinProvider). Regenerates bamlutils/embed.go.

Part of #199 item 4.
